Move SurgeHealthCheck to the surge.internal.health package instead of surge.kafka.streams
