+++ modules/common/src/main/scala/surge/internal/akka/cluster/Shard.scala
@@ -11,8 +11,8 @@ import java.net.URLEncoder
 import org.slf4j.{ Logger, LoggerFactory }
 import surge.akka.cluster.{ Passivate, PerShardLogicProvider }
 import surge.internal.akka.ActorWithTracing
-import surge.kafka.streams.HealthyActor.GetHealth
-import surge.kafka.streams.{ HealthCheck, HealthCheckStatus }
+import surge.internal.health.{ HealthCheck, HealthCheckStatus }
+import surge.internal.health.HealthyActor.GetHealth
 import surge.internal.tracing.TracedMessage
 
 import scala.concurrent.ExecutionContext.Implicits.global
