+++ modules/common/src/main/scala/surge/kafka/streams/AggregateStateStoreKafkaStreams.scala
@@ -9,7 +9,9 @@ import com.typesafe.config.{ Config, ConfigFactory }
 import org.apache.kafka.streams.Topology
 import surge.core.{ Ack, Controllable }
 import surge.health.HealthSignalBusTrait
+import surge.internal.akka.actor.ActorLifecycleManagerActor
 import surge.internal.config.{ BackoffConfig, TimeoutConfig }
+import surge.internal.health.{ HealthCheck, HealthyActor, HealthyComponent }
 import surge.internal.utils.{ BackoffChildActorTerminationWatcher, Logging }
 import surge.kafka.KafkaTopic
 import surge.metrics.Metrics
@@ -115,7 +117,21 @@ class AggregateStateStoreKafkaStreams[Agg >: Null](
           randomFactor = BackoffConfig.StateStoreKafkaStreamActor.randomFactor)
         .withMaxNrOfRetries(BackoffConfig.StateStoreKafkaStreamActor.maxRetries))
 
-    system.actorOf(BackoffChildActorTerminationWatcher.props(system.actorOf(underlyingActorProps), () => onMaxRetries()))
+    val underlyingCreatedActor =
+      ActorLifecycleManagerActor
+        .manage(
+          system,
+          underlyingActorProps,
+          componentName = s"state-store-kafka-streams-$aggregateName",
+          managedActorName = None,
+          startMessageAdapter = Some(() => KafkaStreamManagerActor.Start),
+          stopMessageAdapter = Some(() => KafkaStreamManagerActor.Stop))
+        .ref
+
+    // FIXME When the KafkaStreamManagerActor dies I don't think the ActorLifecycleManagerActor stops with it, so this termination watcher won't be triggered
+    system.actorOf(BackoffChildActorTerminationWatcher.props(underlyingCreatedActor, () => onMaxRetries()))
+
+    underlyingCreatedActor
   }
 
   private[streams] def getTopology: Future[Topology] = {
@@ -156,7 +172,7 @@ class AggregateStateStoreKafkaStreams[Agg >: Null](
   override val controllable: Controllable = new Controllable {
     override def start(): Future[Ack] = {
       implicit val ec: ExecutionContext = system.dispatcher
-      underlyingActor.ask(KafkaStreamManagerActor.Start).mapTo[Ack].andThen(registrationCallback())
+      underlyingActor.ask(ActorLifecycleManagerActor.Start).mapTo[Ack].andThen(registrationCallback())
     }
 
     override def restart(): Future[Ack] = {
@@ -171,7 +187,7 @@ class AggregateStateStoreKafkaStreams[Agg >: Null](
 
     override def stop(): Future[Ack] = {
       implicit val ec: ExecutionContext = system.dispatcher
-      underlyingActor.ask(KafkaStreamManagerActor.Stop).mapTo[Ack].andThen(unregistrationCallback())
+      underlyingActor.ask(ActorLifecycleManagerActor.Stop).mapTo[Ack].andThen(unregistrationCallback())
     }
 
     override def shutdown(): Future[Ack] = stop()
